backend/requirements.txt
asgiref==3.9.1
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
coreapi==2.3.3
coreschema==0.0.4
cryptography==45.0.5
defusedxml==0.7.1
Django==3.2.3
django-filter==23.5
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-extra-fields==3.7.0
filetype==1.2.0
flake8==7.3.0
gunicorn==23.0.0
idna==3.10
itypes==1.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mccabe==0.7.0
oauthlib==3.3.1
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.10
pycodestyle==2.14.0
pycparser==2.22
pyflakes==3.4.0
PyJWT==2.10.1
python-dotenv==1.1.1
python3-openid==3.2.0
pytz==2025.2
requests==2.32.4
requests-oauthlib==2.0.0
setuptools==80.9.0
six==1.17.0
social-auth-app-django==4.0.0
social-auth-core==4.7.0
sqlparse==0.5.3
uritemplate==4.2.0
urllib3==2.5.0
